% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaBinomial.R
\name{BetaBinomial}
\alias{BetaBinomial}
\title{Calculates conjugate Beta posterior from Beta prior and Binomial data}
\usage{
BetaBinomial(y = numeric(), n = 10, a0 = 0.5, b0 = 0.5)
}
\arguments{
\item{y}{An integer data value, the number of successes in n iid Bernoulli trials. Can be empty.}

\item{n}{}

\item{a0}{A positive number, the Beta prior first shape parameter.}

\item{b0}{A postive number, the Beta prior second shape parameter.}
}
\value{
out A list with element a1 and b1, the shape parameters for the posterior Beta distribution.
}
\description{
Calculates conjugate Beta posterior from Beta prior and Binomial data
}
\examples{
BetaBernoulli(y = numeric(), n=100, a0 = 0.5, b0 = 0.5)
BetaBernoulli(y = 4, n=10  a0 = 0.5, b0 = 0.5)
}
