% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaExponential.R
\name{GammaExponential}
\alias{GammaExponential}
\title{Calculates conjugate Gamma posterior from Gamma prior and Exponential data}
\usage{
GammaExponential(y = numeric(), a0 = 2, b0 = 1)
}
\arguments{
\item{y}{A vector of positive numeric data values, can be empty.}

\item{a0}{A positive number, the Gamma prior shape parameter.}

\item{b0}{A positive number, the Gamma prior rate parameter (=1/scale parameter).}
}
\value{
out A list with element a1 and b1, the shape parameters for the posterior Gamma distribution.
}
\description{
Calculates conjugate Gamma posterior from Gamma prior and Exponential data
}
\examples{
GammaExponential(y = numeric(), a0 = 2, b0 = 1)
GammaExponential(y = c(1.215, 3.915, 2.519, 2.593, 6.924), a0 = 2, b0 = 1.0)
}
