% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaPoisson.R
\name{GammaPoisson}
\alias{GammaPoisson}
\title{Calculates conjugate Gamma posterior from Gamma prior and Poisson data}
\usage{
GammaPoisson(y = numeric(), a0 = 2, b0 = 1)
}
\arguments{
\item{y}{A vector of non-negative integer count data, can be empty.}

\item{a0}{A positive number, the Gamma prior shape parameter.}

\item{b0}{A positive number, the Gamma prior rate parameter (=1/scale parameter).}
}
\value{
out A list with element a1 and b1, the shape parameters for the posterior Gamma distribution.
}
\description{
Calculates conjugate Gamma posterior from Gamma prior and Poisson data
}
\examples{
GammaPoisson(y = numeric(), a0 = 2, b0 = 1)
GammaPoisson(y = c(1, 3, 5, 6, 5), a0 = 2, b0 = 1.0)
}
