% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalNormal.R
\name{NormalNormal}
\alias{NormalNormal}
\title{Calculates conjugate Normal posterior from Normal prior and Normal data}
\usage{
NormalNormal(y = numeric(), mu0 = 0, tau0 = 1, sigma = 1)
}
\arguments{
\item{y}{A vector of numeric data, can be empty.}

\item{mu0}{A real number, the Normal prior mean parameter.}

\item{tau0}{A positive number, the Normal prior standard (std) deviation parameter.}

\item{sigma}{A positive number, std deviation of the Normal data distribution, assumed known.}
}
\value{
out A list with element mu1 and tau1, the mean and std deviation parameters for the posterior Normal distribution.
}
\description{
Calculates conjugate Normal posterior from Normal prior and Normal data
}
\examples{
NormalNormal(y = numeric(), mu0 = 2, tau0 = 1, sigma = 2)
NormalNormal(y = c(-1.807, 1.043, -1.700, 2.446, 2.367),  mu0 = 1, tau0 = 1, sigma = 2)
}
